+++ BioViewer/Metal/Shaders/ShadowShader.metal
@@ -20,7 +20,7 @@ struct ShadowVertexOut{
     float4 position [[position]];
     float3 atomCenter [[attribute(0)]];
     half2 billboardMapping [[attribute(1)]];
-    float atom_radius [[attribute(2)]];
+    half atom_radius [[attribute(2)]];
 };
 
 // MARK: - Build constants
